refactor: Moves ETL code from datasets to etl Moves the transformation code in datasets.issues to etl.github
