+++ analytics/src/analytics/etl/github.py
@@ -3,11 +3,18 @@
 from __future__ import annotations
 
 import logging
+from dataclasses import dataclass
 from pathlib import Path
 
-from pydantic import BaseModel
+import pandas as pd
+from pydantic import BaseModel, ValidationError
 
-from analytics.datasets.issues import GitHubIssues, InputFiles
+from analytics.datasets.issues import (
+    GitHubIssues,
+    IssueMetadata,
+    IssueType,
+)
+from analytics.datasets.utils import load_json_file
 from analytics.integrations import github
 
 logger = logging.getLogger(__name__)
@@ -15,6 +22,14 @@ logger = logging.getLogger(__name__)
 # ========================================================
 # Config interfaces
 # ========================================================
+
+
+@dataclass
+class InputFiles:
+    """Expected input files for loading this dataset."""
+
+    roadmap: str
+    sprint: str
 
 
 class GitHubProjectConfig(BaseModel):
@@ -84,7 +99,7 @@ class GitHubProjectETL:
         self.config = config
         # Declare private attributes shared across ETL steps
         self._transient_files: list[InputFiles]
-        self._dataset: GitHubIssues
+        self.dataset: GitHubIssues
 
     def run(self) -> None:
         """Run the ETL pipeline."""
@@ -126,11 +141,15 @@ class GitHubProjectETL:
 
     def transform(self) -> None:
         """Run the transformation step of the ETL pipeline."""
-        self._dataset = GitHubIssues.load_from_json_files(self._transient_files)
+        # Load sprint and roadmap data
+        issues = []
+        for f in self._transient_files:
+            issues.extend(run_transformation_pipeline(files=f))
+        self.dataset = GitHubIssues(pd.DataFrame(data=issues))
 
     def load(self) -> None:
         """Run the load step of the ETL pipeline."""
-        self._dataset.to_json(self.config.output_file)
+        self.dataset.to_json(self.config.output_file)
 
     def _export_roadmap_data(
         self,
@@ -171,3 +190,123 @@ class GitHubProjectETL:
             points_field=sprint_board.points_field,
             output_file=output_file,
         )
+
+
+# ===============================================================
+# Transformation helper functions
+# ===============================================================
+
+
+def run_transformation_pipeline(files: InputFiles) -> list[dict]:
+    """Load data from input files and apply transformations."""
+    # Load sprint and roadmap data
+    sprint_data_in = load_json_file(files.sprint)
+    roadmap_data_in = load_json_file(files.roadmap)
+    # Populate a lookup table with this data
+    lookup: dict = {}
+    lookup = populate_issue_lookup_table(lookup, roadmap_data_in)
+    lookup = populate_issue_lookup_table(lookup, sprint_data_in)
+    # Flatten and write issue level data to output file
+    return flatten_issue_data(lookup)
+
+
+def populate_issue_lookup_table(
+    lookup: dict[str, IssueMetadata],
+    issues: list[dict],
+) -> dict[str, IssueMetadata]:
+    """Populate a lookup table that maps issue URLs to their issue type and parent."""
+    for i, issue in enumerate(issues):
+        try:
+            entry = IssueMetadata.model_validate(issue)
+        except ValidationError as err:  # noqa: PERF203
+            logger.error("Error parsing row %d, skipped.", i)  # noqa: TRY400
+            logger.debug("Error: %s", err)
+            continue
+        lookup[entry.issue_url] = entry
+    return lookup
+
+
+def get_parent_with_type(
+    child_url: str,
+    lookup: dict[str, IssueMetadata],
+    type_wanted: IssueType,
+) -> IssueMetadata | None:
+    """
+    Traverse the lookup table to find an issue's parent with a specific type.
+
+    This is useful if we have multiple nested issues, and we want to find the
+    top level deliverable or epic that a given task or bug is related to.
+    """
+    # Get the initial child issue and its parent (if applicable) from the URL
+    child = lookup.get(child_url)
+    if not child:
+        err = f"Lookup doesn't contain issue with url: {child_url}"
+        raise ValueError(err)
+    if not child.issue_parent:
+        return None
+
+    # Travel up the issue hierarchy until we:
+    #  - Find a parent issue with the desired type
+    #  - Get to an issue without a parent
+    #  - Have traversed 5 issues (breaks out of issue cycles)
+    max_traversal = 5
+    parent_url = child.issue_parent
+    for _ in range(max_traversal):
+        parent = lookup.get(parent_url)
+        # If no parent is found, return None
+        if not parent:
+            return None
+        # If the parent matches the desired type, return it
+        if IssueType(parent.issue_type) == type_wanted:
+            return parent
+        # If the parent doesn't have a its own parent, return None
+        if not parent.issue_parent:
+            return None
+        # Otherwise update the parent_url to "grandparent" and continue
+        parent_url = parent.issue_parent
+
+    # Return the URL of the parent deliverable (or None)
+    return None
+
+
+def flatten_issue_data(lookup: dict[str, IssueMetadata]) -> list[dict]:
+    """Flatten issue data and inherit data from parent epic an deliverable."""
+    result: list[dict] = []
+    for issue in lookup.values():
+        # If the issue is a deliverable or epic, move to the next one
+        if IssueType(issue.issue_type) in [IssueType.DELIVERABLE, IssueType.EPIC]:
+            continue
+
+        # Get the parent deliverable, if the issue has one
+        deliverable = get_parent_with_type(
+            child_url=issue.issue_url,
+            lookup=lookup,
+            type_wanted=IssueType.DELIVERABLE,
+        )
+        if deliverable:
+            # Set deliverable metadata
+            issue.deliverable_title = deliverable.issue_title
+            issue.deliverable_url = deliverable.issue_url
+            issue.deliverable_pillar = deliverable.deliverable_pillar
+            # Set quad metadata
+            issue.quad_id = deliverable.quad_id
+            issue.quad_name = deliverable.quad_name
+            issue.quad_start = deliverable.quad_start
+            issue.quad_end = deliverable.quad_end
+            issue.quad_length = deliverable.quad_length
+
+        # Get the parent epic, if the issue has one
+        epic = get_parent_with_type(
+            child_url=issue.issue_url,
+            lookup=lookup,
+            type_wanted=IssueType.EPIC,
+        )
+        if epic:
+            issue.epic_title = epic.issue_title
+            issue.epic_url = epic.issue_url
+
+        # Add the issue to the results
+        result.append(issue.__dict__)
+
+    # Return the results
+    return result
