+++ analytics/tests/etl/test_github.py
@@ -6,13 +6,18 @@ from pathlib import Path
 from unittest.mock import MagicMock
 
 import pytest
-from analytics.datasets.issues import GitHubIssues, InputFiles
+from analytics.datasets.issues import IssueType
+from analytics.datasets.utils import dump_to_json
 from analytics.etl.github import (
     GitHubProjectConfig,
     GitHubProjectETL,
+    InputFiles,
     RoadmapConfig,
     SprintBoardConfig,
+    get_parent_with_type,
 )
+
+from tests.conftest import issue
 
 
 @pytest.fixture(name="config")
@@ -67,29 +72,57 @@ def test_extract(monkeypatch: pytest.MonkeyPatch, etl: GitHubProjectETL):
     )
 
 
-def test_transform(monkeypatch: pytest.MonkeyPatch, etl: GitHubProjectETL):
+def test_transform(tmp_path: Path, etl: GitHubProjectETL):
     """Test the transform step by mocking GitHubIssues.load_from_json_files."""
-    mock_load_from_json_files = MagicMock(return_value="mock_dataset")
-    monkeypatch.setattr(GitHubIssues, "load_from_json_files", mock_load_from_json_files)
-
-    # Provide a sample transient file to `etl`
-    etl._transient_files = [InputFiles(roadmap="roadmap.json", sprint="sprint.json")]
-
-    # Run the transform method
+    # Arrange - create dummy sprint data
+    sprint_file = str(tmp_path / "sprint-data.json")
+    sprint_data = [
+        issue(issue=1, kind=IssueType.TASK, parent="Epic3", points=2),
+        issue(issue=2, kind=IssueType.TASK, parent="Epic4", points=1),
+    ]
+    roadmap_data = [i.model_dump() for i in sprint_data]
+    dump_to_json(sprint_file, roadmap_data)
+    # Act - create dummy roadmap data
+    roadmap_file = str(tmp_path / "roadmap-data.json")
+    roadmap_data = [
+        issue(issue=3, kind=IssueType.EPIC, parent="Deliverable5"),
+        issue(issue=4, kind=IssueType.EPIC, parent="Deliverable6"),
+        issue(issue=5, kind=IssueType.DELIVERABLE, quad="quad1"),
+    ]
+    roadmap_data = [i.model_dump() for i in roadmap_data]
+    dump_to_json(roadmap_file, roadmap_data)
+    # Arrange
+    output_data = [
+        issue(
+            issue=1,
+            points=2,
+            parent="Epic3",
+            deliverable="Deliverable5",
+            quad="quad1",
+            epic="Epic3",
+        ),
+        issue(
+            issue=2,
+            points=1,
+            parent="Epic4",
+            deliverable=None,
+            quad=None,
+            epic="Epic4",
+        ),
+    ]
+    wanted = [i.model_dump() for i in output_data]
+    etl._transient_files = [InputFiles(roadmap=roadmap_file, sprint=sprint_file)]
+    # Act
     etl.transform()
-
-    # Check if load_from_json_files was called with correct files
-    mock_load_from_json_files.assert_called_once_with(etl._transient_files)
-
-    # Verify that the dataset was assigned correctly
-    assert etl._dataset == "mock_dataset"
+    # Assert
+    assert etl.dataset.to_dict() == wanted
 
 
 def test_load(etl: GitHubProjectETL):
     """Test the load step by mocking the to_json method."""
     mock_to_json = MagicMock()
-    etl._dataset = MagicMock()
-    etl._dataset.to_json = mock_to_json
+    etl.dataset = MagicMock()
+    etl.dataset.to_json = mock_to_json
 
     # Run the load method
     etl.load()
@@ -115,3 +148,104 @@ def test_run(monkeypatch: pytest.MonkeyPatch, etl: GitHubProjectETL):
     mock_extract.assert_called_once()
     mock_transform.assert_called_once()
     mock_load.assert_called_once()
+
+
+class TestGetParentWithType:
+    """Test the get_parent_with_type() method."""
+
+    def test_return_epic_that_is_direct_parent_of_issue(self):
+        """Return the correct epic for an issue that is one level down."""
+        # Arrange
+        task = "Task1"
+        epic = "Epic1"
+        lookup = {
+            task: issue(issue=1, kind=IssueType.TASK, parent=epic),
+            epic: issue(issue=2, kind=IssueType.EPIC, parent=None),
+        }
+        wanted = lookup[epic]
+        # Act
+        got = get_parent_with_type(
+            child_url=task,
+            lookup=lookup,
+            type_wanted=IssueType.EPIC,
+        )
+        # Assert
+        assert got == wanted
+
+    def test_return_correct_deliverable_that_is_grandparent_of_issue(self):
+        """Return the correct deliverable for an issue that is two levels down."""
+        # Arrange
+        task = "Task1"
+        epic = "Epic2"
+        deliverable = "Deliverable3"
+        lookup = {
+            task: issue(issue=1, kind=IssueType.TASK, parent=epic),
+            epic: issue(issue=2, kind=IssueType.EPIC, parent=deliverable),
+            deliverable: issue(issue=3, kind=IssueType.DELIVERABLE, parent=None),
+        }
+        wanted = lookup[deliverable]
+        # Act
+        got = get_parent_with_type(
+            child_url=task,
+            lookup=lookup,
+            type_wanted=IssueType.DELIVERABLE,
+        )
+        # Assert
+        assert got == wanted
+
+    def test_return_none_if_issue_has_no_parent(self):
+        """Return None if the input issue has no parent."""
+        # Arrange
+        task = "task"
+        lookup = {
+            task: issue(issue=1, kind=IssueType.TASK, parent=None),
+        }
+        wanted = None
+        # Act
+        got = get_parent_with_type(
+            child_url=task,
+            lookup=lookup,
+            type_wanted=IssueType.DELIVERABLE,
+        )
+        # Assert
+        assert got == wanted
+
+    def test_return_none_if_parents_form_a_cycle(self):
+        """Return None if the issue hierarchy forms a cycle."""
+        # Arrange
+        task = "Task1"
+        parent = "Task2"
+        lookup = {
+            task: issue(issue=1, kind=IssueType.TASK, parent="parent"),
+            parent: issue(issue=2, kind=IssueType.TASK, parent=task),
+        }
+        wanted = None
+        # Act
+        got = get_parent_with_type(
+            child_url=task,
+            lookup=lookup,
+            type_wanted=IssueType.DELIVERABLE,
+        )
+        # Assert
+        assert got == wanted
+
+    def test_return_none_if_deliverable_is_not_found_in_parents(self):
+        """Return None if the desired type (e.g. epic) isn't found in the list of parents."""
+        # Arrange
+        task = "Task1"
+        parent = "Task2"
+        epic = "Epic3"
+        lookup = {
+            task: issue(issue=1, kind=IssueType.TASK, parent=parent),
+            parent: issue(issue=2, kind=IssueType.TASK, parent=epic),
+            epic: issue(issue=3, kind=IssueType.EPIC, parent=task),
+        }
+        wanted = None
+        # Act
+        got = get_parent_with_type(
+            child_url=task,
+            lookup=lookup,
+            type_wanted=IssueType.DELIVERABLE,
+        )
+        # Assert
+        assert got == wanted
